+++ src/Ema.hs
@@ -7,7 +7,6 @@ import Ema.App as X
 import Ema.Asset as X
 import Ema.Class as X
 import Ema.Route as X
-import Ema.Route.Slug as X
 import Ema.Server as X
   ( emaErrorHtmlResponse,
   )